{"schema":"decodex/commit/1","summary":"Split quantitative product manifests","authority":"manual"}
